feat: production grade implementation of the agent
After the POC phase and creating a proper API spec the implementation
had to be re-done in order to match the API spec and follow general best
practices. This commit does that for the pcap-agent. Additionally there
are minor changes to the API spec where it made sense and the CI scripts
have been adapted as well.

Co-authored-by: Tamara Boehm <tamara.boehm@sap.com>