Move bosh resolver test to separate package to avoid circular dependencies
Bosh resolver tests are mock heavy. If mocks should not be in the resulting binary,
they need to be in a separate package but require lots of functions from the `pcap` package.

Because the tests for `pcap` are also in package `pcap`, importing `test` that contained the
mocks (and imports `pcap`) becomes a circular dependency.