fix: allow agent to shut down right away
To be able to notice that the context has been cancelled we rely on at
least one more packet being read. In practice this should be fine but
especially during testing it can cause the agent to hang because no more
data is captured. This commit closes the pcap handle early to unblock
the read call and allow the agent to shut down right away.

Co-Authored-By: Dominik Froehlich <dominik.froehlich@sap.com>