fix: allow agent to shut down quickly
To be able to notice that the context has been canceled we rely on at
least one more packet being read. In practice this should be fine but
especially during testing it can cause the agent to hang because no more
data is captured. This commit configures a timeout of one second for
reading a single packet. This ensures that we check at least once per
second if the capture has been stopped and allow the agent to stop
capturing even if no further packets are captured.

Co-Authored-By: Dominik Froehlich <dominik.froehlich@sap.com>