fix: untangle shut down logic in api
The API has the processing of messages received from the agent and the
sending of control messages to the agent in one function
`readMsgFromStream` which causes concurrency issues because we rely on
the blocking read call from becoming unblocked in order to tell the
agent to shut down. This, however, prevents the agent from ever shutting
down essentially making it impossible to stop a capture unless there is
a constant flow of packets.

This commit separates the upstream and downstream flows by creating two
separate routines when starting a capture on an agent. It also fixes
smaller issues like misleading names and sending to closed channels.