fix: improve shut down logic (#89)
To be able to notice that the context has been canceled we rely on at
least one more packet being read. In practice this should be fine but
especially during testing it can cause the agent to hang because no more
data is captured. This commit configures a timeout of one second for
reading a single packet. This ensures that we check at least once per
second if the capture has been stopped and allow the agent to stop
capturing even if no further packets are captured.

The API has the processing of messages received from the agent and the
sending of control messages to the agent in one function
`readMsgFromStream` which causes concurrency issues because we rely on
the blocking read call from becoming unblocked in order to tell the
agent to shut down. This, however, prevents the agent from ever shutting
down essentially making it impossible to stop a capture unless there is
a constant flow of packets.

This commit separates the upstream and downstream flows by creating two
separate routines when starting a capture on an agent. It also fixes
smaller issues like misleading names and sending to closed channels.

Co-Authored-By: Dominik Froehlich <dominik.froehlich@sap.com>