ci: de-couple unit and acceptance tests
Concourse is not the best when it comes to PR validations. To reduce the
amount of stuck builds and PRs this commit de-couples the two checks.
This might result in more wasted builds on the acceptance tests which
are expensive, but the current situation which requires manually
triggering the checks is much worse.